fix k8s controller function
