+++ tools/k8s_ns.go
@@ -1,0 +1,23 @@
+package tools
+
+import (
+	apiv1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+//CreateNamespace create namespace
+func CreateNamespace(namespace string) error {
+	client := instanceSingleK8sClient()
+	_, err := client.CoreV1().Namespaces().Get(namespace, metav1.GetOptions{})
+	if err != nil {
+		_, err = client.CoreV1().Namespaces().Create(&apiv1.Namespace{
+			ObjectMeta: metav1.ObjectMeta{
+				Name: namespace,
+			},
+		})
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
